+++ common/snowflake/src/main/java/com/eeminsu/logscope/snowflake/Snowflake.java
@@ -1,4 +1,4 @@
-package kuke.board.common.snowflake;
+package com.eeminsu.logscope.snowflake;
 
 import java.util.random.RandomGenerator;
